refactor: remove runPostBundleDecorators and use directly in bundle
